+++ packages/better-write-localisation/src/en-US/about.ts
@@ -1,10 +1,24 @@
 export default {
   initial: {
-    title: 'Crie como quiser, para o que quiser.',
+    title: 'Expand your horizons.',
     description:
-      'betterwrite.io foi projetado para extrair o máximo possível da criatividade de escritores e entusiastas da escrita.',
+      'betterwrite.io is designed to extract as much creativity from writers and writing enthusiasts as possible.',
   },
-  about: {
-    title: '',
+  entity: {
+    title: 'Control everything absolutely.',
+    description:
+      'Insert titles, paragraphs, line breaks, page breaks, lists, checkboxes, images and drawings separately.',
+    bw: 'Better Write',
+    canvas: 'Others',
+  },
+  portability: {
+    title: 'Create how you want, for what you want.',
+    description:
+      'Convert your designs from other tools and transform them however you like.',
+  },
+  finish: {
+    title1: 'Explore.',
+    title2: 'Discovery.',
+    title3: 'Transform.',
   },
 }